// fe/src/app/api/admin/model/inventoryResponse.ts
/**
 * Shop Admin API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface InventoryResponse { 
    id?: number;
    variantId?: number;
    stock?: number;
    reservedQuantity?: number;
    availableQuantity?: number;
    importPrice?: number;
    sellingPrice?: number;
    color?: string;
    size?: string;
    productName?: string;
    thumbnail?: string;
}